refactor: extract hosted app spi into plugin api module
