+++ platform-plugin-api/src/main/kotlin/io/github/rygel/outerstellar/platform/plugin/HostedAppContribution.kt
