+++ platform-plugin-api/src/main/kotlin/io/github/rygel/outerstellar/platform/web/AdminSummaryCard.kt
