+++ platform-plugin-api/src/main/kotlin/io/github/rygel/outerstellar/platform/web/PlatformPlugin.kt
@@ -16,9 +16,25 @@ typealias PluginLayoutRenderer = io.github.rygel.outerstellar.platform.plugin.Pl
 
 typealias PluginOptions = io.github.rygel.outerstellar.platform.plugin.PluginOptions
 
+typealias PluginAppInfo = io.github.rygel.outerstellar.platform.plugin.PluginAppInfo
+
 typealias HostedAppContext = io.github.rygel.outerstellar.platform.plugin.HostedAppContext
 
 typealias PluginContext = io.github.rygel.outerstellar.platform.plugin.PluginContext
+
+typealias PluginUsers = io.github.rygel.outerstellar.platform.plugin.PluginUsers
+
+typealias PluginAnalytics = io.github.rygel.outerstellar.platform.plugin.PluginAnalytics
+
+typealias PluginNotifications = io.github.rygel.outerstellar.platform.plugin.PluginNotifications
+
+typealias PluginRendering = io.github.rygel.outerstellar.platform.plugin.PluginRendering
+
+typealias PluginApiKeys = io.github.rygel.outerstellar.platform.plugin.PluginApiKeys
+
+typealias PluginOAuth = io.github.rygel.outerstellar.platform.plugin.PluginOAuth
+
+typealias PluginSecurity = io.github.rygel.outerstellar.platform.plugin.PluginSecurity
 
 typealias HostedApp = io.github.rygel.outerstellar.platform.plugin.HostedApp
