+++ platform-plugin-api/src/test/kotlin/io/github/rygel/outerstellar/platform/plugin/HostedAppApiCompatibilityTest.kt
